HTTP validation is different to staleness
Validation and freshness are distinct concerns for HTTP caches, as
defined in RFC7234. Separate out the decorators for Cache-Control and
ETag.
